Update ControlPacket payload type to float and enhance calibration handling in main loop

+++ include/packets.h
@@ -93,7 +93,7 @@ typedef struct __attribute__((packed)) {
     RequestType request_idx;    // request type
     uint8_t flags;              // bitmask: ACK/BUSY/ERR
     uint8_t error_code;         // optional: 0 = none, non-zero = specific error
-    uint32_t payload;
+    float payload;
 } ControlPacket;
